# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Quick Quotes Zimbabwe API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile:
    get:
      operationId: getProfile
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/professional:
    get:
      operationId: getProfessionalProfile
      tags: [users]
      summary: Get professional profile for current user
      responses:
        "200":
          description: Professional profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Professional"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: createProfessionalProfile
      tags: [users]
      summary: Create professional profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProfessionalRequest"
      responses:
        "201":
          description: Professional profile created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Professional"

    patch:
      operationId: updateProfessionalProfile
      tags: [users]
      summary: Update professional profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfessionalRequest"
      responses:
        "200":
          description: Updated professional profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Professional"
...
  /reviews:
    post:
      operationId: createReview
      tags: [reviews]
      summary: Customer submits a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewRequest"
      responses:
        "201":
          description: Review submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/professional/{professionalId}:
    get:
      operationId: getProfessionalReviews
      tags: [reviews]
      summary: Get reviews for a professional
      parameters:
        - name: professionalId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
        suspended:
          type: boolean
        createdAt:
          type: string
        professional:
          $ref: "#/components/schemas/Professional"
      required:
        - id
        - name
        - email
        - role
        - suspended
        - createdAt
    Professional:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        services:
          type: array
          items:
            type: string
        rating:
          type: ["number", "null"]
        verified:
          type: boolean
        completedJobs:
          type: integer
        bio:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
      required:
        - id
        - userId
        - services
        - verified
        - completedJobs
...
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
    CreateProfessionalRequest:
      type: object
      properties:
        services:
          type: array
          items:
            type: string
        bio:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
      required:
        - services
    UpdateProfessionalRequest:
      type: object
      properties:
        services:
          type: array
          items:
            type: string
        bio:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
...
    Review:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        customerId:
          type: integer
        professionalId:
          type: integer
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
        customerName:
          type: ["string", "null"]
      required:
        - id
        - jobId
        - customerId
        - professionalId
        - rating
        - createdAt
    CreateReviewRequest:
      type: object
      properties:
        jobId:
          type: integer
        professionalId:
          type: integer
        rating:
          type: integer
        comment:
          type: ["string", "null"]
      required:
        - jobId
        - professionalId
        - rating