# lib/api-zod/src/generated/types/createReviewRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quick Quotes Zimbabwe API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateReviewRequest {
  jobId: number;
  professionalId: number;
  rating: number;
  /** @nullable */
  comment?: string | null;
}